refactor(ui): split host kit widgets
