+++ src/ui/mod.rs
@@ -1,9 +1,15 @@
+pub mod button;
 pub mod focus;
 pub mod hints;
+pub mod labels;
 pub mod list;
 pub mod overlay;
+pub mod row;
+pub mod shortcuts;
 pub mod sidebar;
 pub mod sidebar_row;
 pub mod style;
+pub mod surface;
+pub mod text_field;
 pub mod theme;
 pub mod widgets;
